Tempus: Fix some warnings-as-errors issues.

+++ packages/tempus/src/Tempus_SolutionHistory_decl.hpp
@@ -99,7 +99,7 @@ public:
     /// Subscript operator
     Teuchos::RCP<SolutionState<Scalar> > operator[](const int i) {
       TEUCHOS_TEST_FOR_EXCEPTION(
-        !((0 <= i) and (i < history_->size())), std::out_of_range,
+        !((0 <= i) and (i < (int)history_->size())), std::out_of_range,
         "Error - SolutionHistory index is out of range.\n"
         << "    [Min, Max] = [ 0, " << history_->size()<< "]\n"
         << "    index = " << i << "\n");
